+++ apps/backend/src/app/api/v1/contact-channels/send-verification-code/route.tsx
@@ -1,7 +1,6 @@
-import * as yup from "yup";
 import { createSmartRouteHandler } from "@/route-handlers/smart-route-handler";
-import { adaptSchema, clientOrHigherAuthTypeSchema, signInEmailSchema, emailVerificationCallbackUrlSchema, yupObject, yupNumber, yupString } from "@stackframe/stack-shared/dist/schema-fields";
 import { KnownErrors } from "@stackframe/stack-shared";
+import { adaptSchema, clientOrHigherAuthTypeSchema, emailVerificationCallbackUrlSchema, signInEmailSchema, yupNumber, yupObject, yupString } from "@stackframe/stack-shared/dist/schema-fields";
 import { contactChannelVerificationCodeHandler } from "../verify/verification-code-handler";
 
 export const POST = createSmartRouteHandler({
@@ -25,7 +24,7 @@ export const POST = createSmartRouteHandler({
     statusCode: yupNumber().oneOf([200]).required(),
     bodyType: yupString().oneOf(["success"]).required(),
   }),
-  async handler({ auth: { project, user }, body: { email, callback_url: callbackUrl } }, fullReq) {
+  async handler({ auth: { project, user }, body: { email, callback_url: callbackUrl } }) {
     if (user.primary_email !== email) {
       throw new KnownErrors.EmailIsNotPrimaryEmail(email, user.primary_email);
     }