+++ packages/stack-shared/src/interface/clientInterface.ts
@@ -9,6 +9,7 @@ import { globalVar } from '../utils/globals';
 import { ReadonlyJson } from '../utils/json';
 import { Result } from "../utils/results";
 import { CurrentUserCrud } from './crud/current-user';
+import { ProviderAccessTokenCrud } from './crud/oauth';
 import { InternalProjectsCrud, ProjectsCrud } from './crud/projects';
 import { TeamPermissionsCrud } from './crud/team-permissions';
 import { TeamsCrud } from './crud/teams';
@@ -610,7 +611,7 @@ export class StackClientInterface {
     url.searchParams.set("error_redirect_url", options.errorRedirectUrl);
 
     if (options.afterCallbackRedirectUrl) {
-      url.searchParams.set("after_callback_redirect_rrl", options.afterCallbackRedirectUrl);
+      url.searchParams.set("after_callback_redirect_url", options.afterCallbackRedirectUrl);
     }
 
     if (options.type === "link") {
@@ -802,13 +803,13 @@ export class StackClientInterface {
     return json;
   }
 
-  async getAccessToken(
+  async createProviderAccessToken(
     provider: string,
     scope: string,
     session: InternalSession,
-  ): Promise<{ accessToken: string }> {
+  ): Promise<ProviderAccessTokenCrud['Client']['Read']> {
     const response = await this.sendClientRequest(
-      `/auth/oauth/connected-account/${provider}/access-token`,
+      `/auth/oauth/connected-accounts/${provider}/access-token`,
       {
         method: "POST",
         headers: {
@@ -818,10 +819,7 @@ export class StackClientInterface {
       },
       session,
     );
-    const json = await response.json();
-    return {
-      accessToken: json.accessToken,
-    };
+    return await response.json();
   }
 
   async createTeamForCurrentUser(