+++ apps/e2e/CHANGELOG.md
@@ -1,5 +1,13 @@
 # e2e-tests
 
+## 2.5.8
+
+### Patch Changes
+
+- Improved docs
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.8
+
 ## 2.5.7
 
 ### Patch Changes